Updating the 'processCommand' function
Now that the function runs, I am adding a fork in order to execute the function. If the user does not enter a terminate command (logout, shutdown, or restart), the fork() sys call will be invoked. The user-input command will be converted to a cString that can then be executed in the "shell" using an execvp() sys call.