Adding functionality to append to an existing file
Previously, writing to any file would occur at the first position. When writing to an existing file, this meant that the contents would be overwritten by the new input. I solved this by adding an if-else statement to check whether the file exists and opening the file with different permissions based on the former condition.